+++ src/main/java/com/rakovets/course/java/core/practice/jcf_set/City.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.jcf_list;
+package com.rakovets.course.java.core.practice.jcf_set;
 
 public enum City {
     Minsk,
