+++ src/main/java/com/rakovets/course/java/core/practice/jcf_set/FineType.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.jcf_list;
+package com.rakovets.course.java.core.practice.jcf_set;
 
 import java.util.Set;
