+++ src/main/java/com/rakovets/course/java/core/practice/jcf_set/Person.java
@@ -1,8 +1,7 @@
-package com.rakovets.course.java.core.practice.jcf_list;
+package com.rakovets.course.java.core.practice.jcf_set;
 
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Set;
 
 public class Person {
     private final String fullName;
